# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MenuCategorySummary {
  name: string;
  itemCount: number;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  isFeatured: boolean;
  /** @nullable */
  averageRating: number | null;
  reviewCount: number;
  createdAt: string;
}

export interface Review {
  id: number;
  menuItemId: number;
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment: string | null;
  createdAt: string;
}

export interface ReviewInput {
  /** @minLength 1 */
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export interface CateringRequest {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  eventType: string;
  eventDate: string;
  guestCount: number;
  itemsRequested: string;
  /** @nullable */
  budgetRange: string | null;
  /** @nullable */
  specialRequests: string | null;
  status: string;
  createdAt: string;
}

export interface CateringRequestInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  eventType: string;
  eventDate: string;
  /** @minimum 1 */
  guestCount: number;
  /** @minLength 1 */
  itemsRequested: string;
  budgetRange?: string;
  specialRequests?: string;
}

export type ListMenuItemsParams = {
category?: string;
featured?: boolean;
};