# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kanz Bakery API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu items and categories
  - name: reviews
    description: Menu item ratings and reviews
  - name: catering
    description: Bulk order and catering requests
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu-categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List menu categories with item counts
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategorySummary"

  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /menu-items/{id}/reviews:
    get:
      operationId: listMenuItemReviews
      tags: [reviews]
      summary: List reviews for a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMenuItemReview
      tags: [reviews]
      summary: Create a review for a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /catering-requests:
    post:
      operationId: createCateringRequest
      tags: [catering]
      summary: Submit a bulk order or catering request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CateringRequestInput"
      responses:
        "201":
          description: Created catering request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CateringRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MenuCategorySummary:
      type: object
      required: [name, itemCount]
      properties:
        name:
          type: string
        itemCount:
          type: integer

    MenuItem:
      type: object
      required:
        [
          id,
          name,
          description,
          price,
          category,
          imageUrl,
          isFeatured,
          averageRating,
          reviewCount,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    Review:
      type: object
      required: [id, menuItemId, authorName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    ReviewInput:
      type: object
      required: [authorName, rating]
      properties:
        authorName:
          type: string
          minLength: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    CateringRequest:
      type: object
      required:
        [
          id,
          fullName,
          email,
          phone,
          eventType,
          eventDate,
          guestCount,
          itemsRequested,
          budgetRange,
          specialRequests,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
          format: date
        guestCount:
          type: integer
        itemsRequested:
          type: string
        budgetRange:
          type: ["string", "null"]
        specialRequests:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    CateringRequestInput:
      type: object
      required:
        [fullName, email, phone, eventType, eventDate, guestCount, itemsRequested]
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
        phone:
          type: string
          minLength: 1
        eventType:
          type: string
          minLength: 1
        eventDate:
          type: string
          format: date
        guestCount:
          type: integer
          minimum: 1
        itemsRequested:
          type: string
          minLength: 1
        budgetRange:
          type: string
        specialRequests:
          type: string