# lib/api-zod/src/generated/types/cateringRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CateringRequest {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  eventType: string;
  eventDate: Date;
  guestCount: number;
  itemsRequested: string;
  /** @nullable */
  budgetRange: string | null;
  /** @nullable */
  specialRequests: string | null;
  status: string;
  createdAt: Date;
}