# lib/api-zod/src/generated/types/cateringRequestInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CateringRequestInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  eventType: string;
  eventDate: Date;
  /** @minimum 1 */
  guestCount: number;
  /** @minLength 1 */
  itemsRequested: string;
  budgetRange?: string;
  specialRequests?: string;
}