# lib/api-zod/src/generated/types/menuCategorySummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuCategorySummary {
  name: string;
  itemCount: number;
}