# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  isFeatured: boolean;
  /** @nullable */
  averageRating: number | null;
  reviewCount: number;
  createdAt: Date;
}