# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  menuItemId: number;
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment: string | null;
  createdAt: Date;
}