# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kanz Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  /** @minLength 1 */
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}